Consolidate AssemblyPath with TestAssembly.
