+++ src/Fixie.TestAdapter/VsTestExecutor.cs
@@ -8,7 +8,6 @@
     using Microsoft.VisualStudio.TestPlatform.ObjectModel;
     using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
     using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
-    using static AssemblyPath;
     using static TestAssembly;
 
     [ExtensionUri(Id)]
